# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A clean-architecture REST API built with Gin and PostgreSQL.",
        "title": "Go Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Abdul Falaq",
            "url": "https://github.com/abdulfalaq5"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/users": {
            "get": {
                "description": "Returns a list of all users ordered by ID ascending.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "List all users",
                "responses": {
                    "200": {
                        "description": "List of users",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/github_com_falaqmsi_go-example_internal_model.User"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                },
                                "meta": {
                                    "$ref": "#/definitions/github_com_falaqmsi_go-example_pkg_response.Meta"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_falaqmsi_go-example_pkg_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Inserts a new user record. Email must be unique.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "User payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_falaqmsi_go-example_internal_model.CreateUserInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/github_com_falaqmsi_go-example_internal_model.User"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "meta": {
                                    "$ref": "#/definitions/github_com_falaqmsi_go-example_pkg_response.Meta"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_falaqmsi_go-example_pkg_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_falaqmsi_go-example_pkg_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "description": "Returns a single user record identified by its numeric ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get a user by ID",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/github_com_falaqmsi_go-example_internal_model.User"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "meta": {
                                    "$ref": "#/definitions/github_com_falaqmsi_go-example_pkg_response.Meta"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_falaqmsi_go-example_pkg_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_falaqmsi_go-example_pkg_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_falaqmsi_go-example_pkg_response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates the name and email of an existing user by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update an existing user",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_falaqmsi_go-example_internal_model.UpdateUserInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User updated",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/github_com_falaqmsi_go-example_internal_model.User"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "meta": {
                                    "$ref": "#/definitions/github_com_falaqmsi_go-example_pkg_response.Meta"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error or invalid ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_falaqmsi_go-example_pkg_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_falaqmsi_go-example_pkg_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_falaqmsi_go-example_pkg_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Permanently removes a user record by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Delete a user",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "meta": {
                                    "$ref": "#/definitions/github_com_falaqmsi_go-example_pkg_response.Meta"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_falaqmsi_go-example_pkg_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_falaqmsi_go-example_pkg_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_falaqmsi_go-example_pkg_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns the current operational status of the API server.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "Server is healthy",
                        "schema": {
                            "$ref": "#/definitions/github_com_falaqmsi_go-example_internal_model.HealthStatus"
                        }
                    },
                    "500": {
                        "description": "Service unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_falaqmsi_go-example_pkg_response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_falaqmsi_go-example_internal_model.CreateUserInput": {
            "description": "Request body for creating a user.",
            "type": "object",
            "required": [
                "email",
                "name"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "alice@example.com"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "Alice"
                }
            }
        },
        "github_com_falaqmsi_go-example_internal_model.HealthStatus": {
            "type": "object",
            "properties": {
                "env": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "github_com_falaqmsi_go-example_internal_model.UpdateUserInput": {
            "description": "Request body for updating a user.",
            "type": "object",
            "required": [
                "email",
                "name"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "alice.updated@example.com"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "Alice Updated"
                }
            }
        },
        "github_com_falaqmsi_go-example_internal_model.User": {
            "description": "Full user record returned by the API.",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-04-13T08:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "alice@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Alice"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-04-13T08:00:00Z"
                }
            }
        },
        "github_com_falaqmsi_go-example_pkg_response.ErrorResponse": {
            "description": "Standard error response wrapper.",
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string",
                    "example": "error description"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_falaqmsi_go-example_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "github_com_falaqmsi_go-example_pkg_response.Meta": {
            "description": "Response metadata always present on every API response.",
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 10
                },
                "timestamp": {
                    "type": "string",
                    "example": "2026-04-13T08:00:00Z"
                },
                "total_rows": {
                    "type": "integer",
                    "example": 100
                }
            }
        }
    }
}